# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastShip API V1",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get all categories",
                "operationId": "cc06c43ee00f3380675b4c51af686dfb",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all categories"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/by_store": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get all categories by store",
                "operationId": "8218dcd093c0bc1aa3478dacdfcd04a2",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "store_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all categories by store"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/create": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create categories",
                "operationId": "dcb36b465bd9bef8d6d9590b7929beef",
                "requestBody": {
                    "description": "Categories object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_vi": {
                                        "description": "Tên VN",
                                        "type": "string",
                                        "example": "Name vi"
                                    },
                                    "name_en": {
                                        "description": "Tên EN",
                                        "type": "string",
                                        "example": "Name en"
                                    },
                                    "name_zh": {
                                        "description": "Tên ZH",
                                        "type": "string",
                                        "example": "Name zh"
                                    },
                                    "name_hu": {
                                        "description": "Tên HU",
                                        "type": "string",
                                        "example": "name hu"
                                    },
                                    "image": {
                                        "description": "Đường dẫn đến hình ảnh của sản phẩm hoặc mã giảm giá.",
                                        "type": "string",
                                        "example": "abcd"
                                    },
                                    "description": {
                                        "description": "Mô tả chi tiết về thể loại",
                                        "type": "string",
                                        "example": "abcd"
                                    },
                                    "arrange": {
                                        "description": "Sắp xếp thẻ loại",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "parent_id": {
                                        "description": "ID thể loại cha. nếu ko có thì để null",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "store_id": {
                                        "description": "ID của cửa hàng.",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create categories Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/update": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Update categories",
                "operationId": "3a922298d862c47ecffc04736fbc85fc",
                "requestBody": {
                    "description": "categories object that needs to be update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID thể loại",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name_vi": {
                                        "description": "Tên VN",
                                        "type": "string",
                                        "example": "Name vi"
                                    },
                                    "name_en": {
                                        "description": "Tên EN",
                                        "type": "string",
                                        "example": "Name en"
                                    },
                                    "name_zh": {
                                        "description": "Tên ZH",
                                        "type": "string",
                                        "example": "Name zh"
                                    },
                                    "name_hu": {
                                        "description": "Tên HU",
                                        "type": "string",
                                        "example": "name hu"
                                    },
                                    "image": {
                                        "description": "Đường dẫn đến hình ảnh của sản phẩm hoặc mã giảm giá.",
                                        "type": "string",
                                        "example": "abcd"
                                    },
                                    "description": {
                                        "description": "Mô tả chi tiết về thể loại",
                                        "type": "string",
                                        "example": "abcd"
                                    },
                                    "arrange": {
                                        "description": "Sắp xếp thẻ loại",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "parent_id": {
                                        "description": "ID thể loại cha. nếu ko có thì để null",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1"
                                    },
                                    "store_id": {
                                        "description": "ID của cửa hàng.",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update categories Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/delete": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete categories",
                "operationId": "8ddd022be274bfe0fa5860d730fca029",
                "requestBody": {
                    "description": "Delete categories",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delete successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}